KGBros/back/requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.2.0
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-seed==0.2.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
Faker==4.1.3
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.7
proto-plus==1.11.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-decouple==3.3
python-telegram-bot==13.0
pytz==2020.4
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
rsa==4.6
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
traitlets==5.0.4
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
